First pass at Turtlesets generated by copy only.
So, I changed my mind. Using object inheritance for creating turtlesets
on .with() is a bad idea because if a turtle is created after the
inheriting set is created it will nonetheless contain that
turtle. Namely

t = turtles.with(@color == color.red)
bob = new Turtle
bob.color = color.blue
t now contains bob. Bad news.

We could fix this with a bound on who-number for each new turtleset,
but, that seems yucky.